+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # ChangeLog
 
-* fix JEI plugin not showing uses or crafts
+* remake how the phantom fluid widget works